Move enum to lib to enable non-CLI usage
